Slight reorganization of application files
