+++ module/applications/_module.mjs
@@ -1,6 +1,6 @@
 export * as actor from "./actor/_module.mjs";
+export * as item from "./item/_module.mjs";
 
-export {default as ItemSheet5e} from "./item-sheet.mjs";
 export {default as ProficiencySelector} from "./proficiency-selector.mjs";
 export {default as PropertyAttribution} from "./property-attribution.mjs";
 export {default as TraitSelector} from "./trait-selector.mjs";
