+++ module/applications/actor/_module.mjs
@@ -4,14 +4,12 @@ export {default as ActorSheet5eNPC} from "./npc-sheet.mjs";
 export {default as ActorSheet5eVehicle} from "./vehicle-sheet.mjs";
 
 export {default as ActorAbilityConfig} from "./ability-config.mjs";
-export {default as AbilityUseDialog} from "./ability-use-dialog.mjs";
-export {default as ActorArmorConfig} from "./actor-armor.mjs"
-export {default as ActorSheetFlags} from "./actor-flags.mjs";
-export {default as ActorTypeConfig} from "./actor-type.mjs"
+export {default as ActorArmorConfig} from "./armor-config.mjs";
 export {default as ActorHitDiceConfig} from "./hit-dice-config.mjs";
 export {default as LongRestDialog} from "./long-rest.mjs";
 export {default as ActorMovementConfig} from "./movement-config.mjs";
 export {default as ActorSensesConfig} from "./senses-config.mjs";
+export {default as ActorSheetFlags} from "./sheet-flags.mjs";
 export {default as ShortRestDialog} from "./short-rest.mjs";
 export {default as ActorSkillConfig} from "./skill-config.mjs";
-
+export {default as ActorTypeConfig} from "./type-config.mjs";
