+++ module/applications/actor/armor-config.mjs
