+++ module/applications/actor/type-config.mjs
