+++ module/applications/item/_module.mjs
@@ -1,0 +1,3 @@
+export {default as ItemSheet5e} from "./item-sheet.mjs";
+
+export {default as AbilityUseDialog} from "./ability-use-dialog.mjs";
