+++ module/applications/item/item-sheet.mjs
@@ -1,7 +1,7 @@
-import AdvancementManager from "../advancement/advancement-manager.mjs";
-import ProficiencySelector from "./proficiency-selector.mjs";
-import TraitSelector from "./trait-selector.mjs";
-import ActiveEffect5e from "../documents/active-effect.mjs";
+import AdvancementManager from "../../advancement/advancement-manager.mjs";
+import ProficiencySelector from "../proficiency-selector.mjs";
+import TraitSelector from "../trait-selector.mjs";
+import ActiveEffect5e from "../../documents/active-effect.mjs";
 
 /**
  * Override and extend the core ItemSheet implementation to handle specific item types.
